Added /canvas route
