+++ src/app/app-routing.module.ts
@@ -1,7 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { CanvasComponent } from './components/canvas/canvas.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'canvas', component: CanvasComponent }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
